+++ src/environments/environment.ts
@@ -1,6 +1,7 @@
 export const environment = {
   production: false,
   certIdentifierUrl: 'https://cgcom.127.0.0.1.nip.io:4443/cert',
+  issuanceStartUrl: '',
   bootstrapApiUrl: 'https://cgcom.127.0.0.1.nip.io:4443/identify/api/bootstrap',
   walletCallbackBase: 'https://cgcom.127.0.0.1.nip.io:4443/wallet/protocol/callback',
   oidcAuthorizationEndpoint: 'https://cgcom.127.0.0.1.nip.io:4443/verifier/oidc/authorize',